+++ pyroomacoustics/bss/__init__.py
@@ -1,3 +1,4 @@
+# encoding: utf-8
 '''
 Blind Source Separation
 =======================
@@ -13,6 +14,10 @@ Trinicon
 ILRMA
     | Independent Low-Rank Matrix Analysis [3]_
     | :py:mod:`pyroomacoustics.bss.ilrma`
+SparseAuxIVA
+    | Sparse Independent Vector Analysis [4]_
+    | :py:mod `pyroomacoustics.bss.sparseauxiva`
+
 
 A few commonly used functions, such as projection back, can be found in
 :py:mod:`pyroomacoustics.bss.common`.
@@ -29,12 +34,17 @@ References
     doi:10.1016/j.sigpro.2005.06.022, 2006.
 
 .. [3] D. Kitamura, N. Ono, H. Sawada, H. Kameoka, and H. Saruwatari
-     *Determined Blind Source Separation with Independent Low-Rank Matrix Analysis*,
-     in Audio Source Separation, S. Makino, Ed. Springer, 2018, pp.  125-156.
+    *Determined Blind Source Separation with Independent Low-Rank Matrix Analysis*,
+    in Audio Source Separation, S. Makino, Ed. Springer, 2018, pp.  125-156.
+
+.. [4] Janský, Jakub & Koldovský, Zbyněk & Ono, Nobutaka. (2016). A computationally
+    cheaper method for blind speech separation based on AuxIVA and incomplete demixing
+    transform. 1-5. 10.1109/IWAENC.2016.7602921.
 
 '''
 
 from .trinicon import trinicon
 from .auxiva import auxiva, f_contrasts
 from .ilrma import ilrma
-from .common import projection_back
+from .sparseauxiva import sparseauxiva
+from .common import projection_back, sparir