+++ CHANGELOG.rst
@@ -11,7 +11,21 @@ adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 `Unreleased`_
 -------------
 
-Nothing yet.
+Added
+~~~~~
+- New algorithm for blind source separation (BSS): Sparse Independent Vector Analysis (SparseAuxIVA)
+
+Changed
+~~~~~~~
+
+
+Bugfix
+~~~~~~
+
+- Remove `np.squeeze` in STFT as it caused errors when an axis that shouldn't
+  be squeezed was equal to 1.
+- `Beamformer.process` was using old (non-existent) STFT function. Changed to
+  using one-shot function from `transform` module.
 
 `0.1.21`_ - 2018-12-20
 ----------------------
@@ -21,8 +35,11 @@ Added
 
 - Adds several options to ``pyroomacoustics.room.Room.simulate`` to finely
   control the SNR of the microphone signals and also return the microphone
-  signals with individual sources, prior to mix (usefull for BSS evaluation)
+  signals with individual sources, prior to mix (useful for BSS evaluation)
 - Add subspace denoising approach in ``pyroomacoustics.denoise.subspace``.
+- Add iterative Wiener filtering approach for single channel denoising in
+  ``pyroomacoustics.denoise.iterative_wiener``.
+
 
 Changed
 ~~~~~~~
@@ -41,6 +58,7 @@ Bugfix
 - Corrects a bug in the update of the demixing matrix in ``pyroomacoustics.bss.auxiva``
 - Corrects invalid memory access in the ``pyroomacoustics.build_rir`` cython accelerator
   and adds a unit test that checks the cython code output is correct
+- Fix bad handling of 1D `b` vectors in ```pyroomacoustics.levinson``.
 
 `0.1.20`_ - 2018-10-04
 ----------------------